+++ src/main/java/com/elduin/weird_mobs/WeirdMobs.java
@@ -1,15 +1,14 @@
 package com.elduin.weird_mobs;
 
+import com.elduin.weird_mobs.entity.ModEntities;
 import com.elduin.weird_mobs.platform.Platform;
+import com.elduin.weird_mobs.platform.fabric.FabricPlatform;
 
 import net.minecraft.resources.ResourceLocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.elduin.weird_mobs.platform.fabric.FabricPlatform;
-
-@SuppressWarnings("LoggingSimilarMessage")
-public class ModTemplate {
+public class WeirdMobs {
 
 	public static final String MOD_ID = /*$ mod_id*/ "weird_mobs";
 	public static final String MOD_VERSION = /*$ mod_version*/ "1.0.0";
@@ -19,13 +18,13 @@ public class ModTemplate {
 	private static final Platform PLATFORM = createPlatformInstance();
 
 	public static void onInitialize() {
-		LOGGER.info("Initializing {} on {}", MOD_ID, ModTemplate.xplat().loader());
-		LOGGER.debug("{}: { version: {}; friendly_name: {} }", MOD_ID, MOD_VERSION, MOD_FRIENDLY_NAME);
+		LOGGER.info("Initializing {} on {}", MOD_ID, WeirdMobs.xplat().loader());
+		ModEntities.register();
 	}
 
 	public static void onInitializeClient() {
-		LOGGER.info("Initializing {} Client on {}", MOD_ID, ModTemplate.xplat().loader());
-		LOGGER.debug("{}: { version: {}; friendly_name: {} }", MOD_ID, MOD_VERSION, MOD_FRIENDLY_NAME);
+		LOGGER.info("Initializing {} Client on {}", MOD_ID, WeirdMobs.xplat().loader());
+		com.elduin.weird_mobs.client.ModClient.register();
 	}
 
 	static Platform xplat() {
@@ -36,11 +35,7 @@ public class ModTemplate {
 		return new FabricPlatform();
 	}
 
-	private static ResourceLocation id(String path) {
+	public static ResourceLocation id(String path) {
 		return ResourceLocation.fromNamespaceAndPath(MOD_ID, path);
-	}
-
-	private static ResourceLocation id(String namespace, String path) {
-		return ResourceLocation.fromNamespaceAndPath(namespace, path);
 	}
 }