+++ src/main/java/com/elduin/weird_mobs/event/ExampleEventHandler.java
@@ -1,11 +1,11 @@
 package com.elduin.weird_mobs.event;
 
-import com.elduin.weird_mobs.ModTemplate;
+import com.elduin.weird_mobs.WeirdMobs;
 import net.minecraft.server.level.ServerPlayer;
 
 public class ExampleEventHandler {
 
 	public static void onPlayerHurt(ServerPlayer player) {
-		ModTemplate.LOGGER.info("{} took damage.", player.getDisplayName());
+		WeirdMobs.LOGGER.info("{} took damage.", player.getDisplayName());
 	}
 }