+++ src/main/java/com/elduin/weird_mobs/platform/fabric/FabricClientEntrypoint.java
@@ -2,7 +2,7 @@ package com.elduin.weird_mobs.platform.fabric;
 
 //? fabric {
 
-import com.elduin.weird_mobs.ModTemplate;
+import com.elduin.weird_mobs.WeirdMobs;
 import dev.kikugie.fletching_table.annotation.fabric.Entrypoint;
 import net.fabricmc.api.ClientModInitializer;
 
@@ -11,7 +11,7 @@ public class FabricClientEntrypoint implements ClientModInitializer {
 
 	@Override
 	public void onInitializeClient() {
-		ModTemplate.onInitializeClient();
+		WeirdMobs.onInitializeClient();
 	}
 
 }