+++ src/main/java/com/elduin/weird_mobs/platform/fabric/FabricEntrypoint.java
@@ -2,7 +2,7 @@ package com.elduin.weird_mobs.platform.fabric;
 
 //? fabric {
 
-import com.elduin.weird_mobs.ModTemplate;
+import com.elduin.weird_mobs.WeirdMobs;
 import dev.kikugie.fletching_table.annotation.fabric.Entrypoint;
 import net.fabricmc.api.ModInitializer;
 
@@ -11,7 +11,7 @@ public class FabricEntrypoint implements ModInitializer {
 
 	@Override
 	public void onInitialize() {
-		ModTemplate.onInitialize();
+		WeirdMobs.onInitialize();
 		FabricEventSubscriber.registerEvents();
 	}
 }